\name{HealthPromotion}
\alias{HealthPromotion}
\docType{data}
\title{}
\description{}
\usage{data(HealthPromotion)}
\format{
  A data frame with 4605 observations on the following 17 variables.
  \describe{
    \item{\code{ID}}{Subject identifier for the 1365 subjects}
    \item{\code{Visit}}{}
    \item{\code{Gender}}{Reported as "MALE" or "FEMALE"}
    \item{\code{Ethnicity}}{"CHINESE", "MALEY","INDIAN","OTHER"}
    \item{\code{Age}}{Age at baseline}
    \item{\code{Occupation}}{Occupation at baseline: "ACADEMIC", "SUPPORT I", "SUPPORT II"}
    \item{\code{Smoking}}{Smoking category at baseline: "NON-SMOKER","EX-SMOKER","SMOKER"}
    \item{\code{BMIGroup}}{BMI group at baseline: "NORMAL", "OVERWEIGHT", "OBESE"}
    \item{\code{HypertensionHist}}{History of hypertension at baseline: "YES" or "NO"}
    \item{\code{HypertensionMed0}}{Hypertension medicine at baseline: "YES" or "NO"}
    \item{\code{HypertensionGroup}}{Hypertension group at baseline: "HEALTHY", "AT RISK", "HYPERTENSIVE"}
    \item{\code{Height}}{Height (m)}
    \item{\code{Weight}}{Weight (kg)}
    \item{\code{BMI}}{BMI (kg/m^2)}
    \item{\code{HypertensionMed}}{Hypertension medicine at visit}
    \item{\code{SystolicBP}}{Systolic blod pressure (mmHg)}
    \item{\code{DiastolicBP}}{Diastolic blod pressure (mmHg)}
  }
}
\references{
Eng, J. Y., Moy, F. M., & Bulgiba, A. (2016). Impact of a workplace health promotion program on employees' blood pressure in a public university. PloS one, 11(2), e0148307.
}

\examples{
data(HealthPromotion)

\dontrun{
}

}


\keyword{datasets}
